Latest Enigma document + the EnigmaKernel.cpp file.
The latest Enigma document contains a Table of Contents, and 3 new
sections: Section 3, describing the Enigma kernel; Section 4, describing
the testing that will be required to verify proper operation of the
Enigma kernel; and Section 5, describing how one tries to 'break' the
Enigma code.  The EnigmaKernel.cpp file is a standalone, bare-bones
Enigma simulator that uses a four (4) letter alphabet (A, B, C & D).  It
is easily expandable to the full 26 letter alphabet.