Here is the latest version of the Enigma Kernel, which fixes the known problem with using the counter variable, creates a forward and reverse
Here is the latest version of the Enigma Kernel, which fixes the known
problem with using the counter variable, creates a forward and reverse
fixed rotor - needed for the Railway Enigma that does not use the
identity mapping on the fixed rotor, and makes it very easy, now, to
specify the notch position.  Many more comments have been added, and
existing ones modified.  Again, the alphabet has been limited to a 4
letter alphabet for ease of testing, but it is easy to extend to the
full 26 letter alphabet.